Hourly electricity load

Hourly tsibble with electricity loads and load forecasts from the ENTSO-E Transparency Platform. The dataset contains time series data from 2019-01-01 00:00:00 to 2019-12-31 23:00:00 for 8 bidding zones within europe (DE, DK1, ES, FI, FR, NL, NO1, SE1). The data are aggregated in time from a 15-minutes interval to an hourly basis.

data(elec_load)

Format

A time series object of class tsibble with 122.640 rows and 5 columns:

Time: Date and time (index variable)

Series: Time series name (key variable)

Unit: Measured unit (key variable)

BZN: Bidding zone (key variable)

Value: Measurement variable

Source

ENTSO-E Transparency Platform

Examples

data(elec_load)