Estimate accuracy metrics to evaluate point forecast

This function estimates several accuracy metrics to evaluate the accuracy of point forecasts. Either along the forecast horizon or along the train-test-splits. By default, the following accuracy metrics are provided:

ME: mean error

MAE: mean absolute error

MSE: mean squared error

RMSE: root mean squared error

MAPE: mean absolute percentage error

sMAPE: scaled mean absolute percentage error

MPE: mean percentage error

MASE: mean absolute scaled error

error_metrics(fcst, test, train = NULL, period = NULL, by = "split")

Arguments

fcst
A fable containing the forecasts for the models, splits, etc.

test
A tsibble containing the testing data.

train
A tsibble containing the training data. If train = NULL, no MASE is calculated.

period
Integer value. The period used for the estimation of the in-sample MAE of seasonal naive forecast. The in-sample MAE is required for scaling the MASE.

by
Character value. Either accuracy is estimated by split or horizon.

Value

A tibble containing the accuracy metrics for all key variables and models.